// src/app/module-manager/models/UserApplicationAcces.ts
export class UserApplicationAcces {
    uuidApplication: string ="";
    codeApplication: string ="";
    nomApplication: string ="";
    urlApplication: string ="";
    descriptionApplication: string ="";
    version: string ="";
    environnement: string ="";
    applicationEnabled: boolean = true;
    color: string ="";
    icon: string = "";
    hasAccess: boolean = true;
    enabled: boolean = true;
    nonLocked: boolean = true;
    nonExpired: boolean = true;
}